chore(dev): refactor thread page
