+++ frontend/src/app/(dashboard)/projects/[projectId]/thread/[threadId]/layout.tsx
@@ -1,0 +1,8 @@
+export default function ThreadLayout({
+    children,
+  }: {
+    children: React.ReactNode;
+  }) {
+    return <>{children}</>;
+  }
+
