+++ frontend/src/app/(dashboard)/projects/[projectId]/thread/_components/index.ts
@@ -1,0 +1,3 @@
+export { ThreadError } from './ThreadError';
+export { UpgradeDialog } from './UpgradeDialog';
+export { ThreadLayout } from './ThreadLayout';
